Split GetTempDir from internal/util
With this pkg/parse does not depend on libimage.

[NO NEW TESTS NEEDED]

Based on Miloslav's work: https://github.com/containers/podman/pull/19718
